HSEARCH-436 Refining metadata API. Moving Type enum into FieldSettingsDescriptor and changing the types to BASIC and NUMERIC.
Created subtype for FieldSettingsDescriptor - NumericFieldSettingsDescriptor - with numeric specific properties. Offering FieldSettingsDescriptor#as for unwrapping

Add PropertyDescriptor#isId

Updated tests, splitting MetadataTest into several descriptor specific test classes
