+++ engine/src/main/java/org/hibernate/search/metadata/FieldDescriptor.java
@@ -34,12 +34,6 @@ import org.hibernate.search.bridge.FieldBridge;
  * @author Hardy Ferentschik
  */
 public interface FieldDescriptor extends FieldSettingsDescriptor {
-
-	/**
-	 * @return the field type for this field
-	 */
-	Type getFieldType();
-
 	/**
 	 * @return the string used to index {@code null} values. {@code null} in case null values are not indexed
 	 */
@@ -61,19 +55,4 @@ public interface FieldDescriptor extends FieldSettingsDescriptor {
 	 * @return the analyzer used for this field, {@code null} if the field is not analyzed
 	 */
 	Analyzer getAnalyzer();
-
-	/**
-	 * Defines different logical field types
-	 */
-	public static enum Type {
-		/**
-		 * This field is used as the document id
-		 */
-		ID,
-
-		/**
-		 * A basic field generated by a field bridge
-		 */
-		BASIC
-	}
 }
