+++ engine/src/main/java/org/hibernate/search/metadata/IndexedTypeDescriptor.java
@@ -77,9 +77,21 @@ public interface IndexedTypeDescriptor extends FieldContributor {
 	Set<PropertyDescriptor> getIndexedProperties();
 
 	/**
+	 * Retrieves the property descriptor for a given property name.
+	 *
+	 * @param propertyName the property name for which to return descriptor, cannot be {@code null}
+	 *
+	 * @return the property descriptor for the specified property name. {@code null} is returned in case a property with
+	 *         the specified name does not exist
+	 *
+	 * @throws IllegalArgumentException in case {@code propertyName} is {@code null}
+	 */
+	PropertyDescriptor getProperty(String propertyName);
+
+	/**
 	 * Retrieves the field descriptor for a given field name.
 	 *
-	 * @param fieldName the field name for which to return descriptor. Cannot be {@code null}
+	 * @param fieldName the field name for which to return descriptor, cannot be {@code null}
 	 *
 	 * @return the field descriptor for the specified field name. {@code null} is returned in case a field with the specified name does not exist
 	 *
@@ -90,10 +102,12 @@ public interface IndexedTypeDescriptor extends FieldContributor {
 	/**
 	 * Returns the set of index descriptors for the indexed field generated by the property with the given name.
 	 *
-	 * @param propertyName the property name
+	 * @param propertyName the property name, cannot be {@code null}
 	 *
 	 * @return the set of index descriptors for the indexed field generated by the property with the given name. The empty set
 	 *         is returned in case the property does not exist or does not produce any indexed fields.
+	 *
+	 * @throws IllegalArgumentException in case {@code propertyName} is {@code null}
 	 */
 	Set<FieldDescriptor> getFieldsForProperty(String propertyName);
 }
