+++ engine/src/main/java/org/hibernate/search/util/logging/impl/Log.java
@@ -712,4 +712,13 @@ public interface Log extends BasicLogger {
 
 	@Message(id = 179, value = "Unable to create a FullTextEntityManager from a null EntityManager")
 	IllegalArgumentException getNullSessionPassedToFullEntityManagerCreationException();
+
+	@Message(id = 180, value = "Unable to cast %s of type %s to %s")
+	ClassCastException getUnableToNarrowFieldDescriptorException(String actualDescriptorType, String type, String expectedType);
+
+	@Message(id = 181, value = "'null' is not a valid property name")
+	IllegalArgumentException getPropertyNameCannotBeNullException();
+
+	@Message(id = 182, value = "'null' is not a valid field name")
+	IllegalArgumentException getFieldNameCannotBeNullException();
 }
